液位传感器检测到水后，延时2s才关闭电磁阀

+++ include/timer.h
@@ -13,9 +13,9 @@
 #include "config.h"
 
 
-static u8 TIMER0_CNT = 0;     // 定时器0计数器1
+static u8 TIMER0_CNT = 0;      // 定时器0计数器1
 static u16 TIMER0_CNT2 = 0;    // 定时器0计数器2
-static u16 TIMER0_CNT3 = 0;    // 定时器0计数器3
+static u8 TIMER0_CNT3 = 0;     // 定时器0计数器3
 
 extern void timer_init();
